backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "abraham",email = "abrahampruebasespeciales@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "boto3 (>=1.40.74,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
